smart: Fix change table walk, widen device_idx, rename migration
- sataChangeByDeviceTable: walk the full table and extract lastChange
  explicitly — table(2) puts the column name at depth 3, so walking
  a single column OID produced {colName: value} arrays instead of scalars
- smart_sata_change.device_idx: unsignedSmallInteger → unsignedInteger
  (SNMP instance indexes are OID-derived 32-bit values, not sequential)
- Rename migration 2026_05_27 → 2026_06_06 to reflect current date

+++ database/migrations/2026_06_06_000000_create_smart_tables.php
@@ -15,7 +15,7 @@ return new class extends Migration
 
         Schema::create('smart_sata_change', function (Blueprint $table) {
             $table->unsignedInteger('app_id');
-            $table->unsignedSmallInteger('device_idx');            // smartmonDeviceIndex
+            $table->unsignedInteger('device_idx');                 // smartmonDeviceIndex
             $table->unsignedTinyInteger('table_id');               // SATA_TID_* constant
             $table->unsignedInteger('subindex')->default(0);       // 0 = device-level; >0 = page/error-entry-level
             $table->string('last_change', 32)->nullable();         // DateAndTime from MIB
